fix sb init
